requirements.txt
Flask==3.0.0
pdfplumber==0.10.3
python-docx==1.1.2
spacy==3.7.4
pandas==2.2.0
python-dateutil==2.9.0
